Rework the incompatible constraint sample
This changes the sample to illustrate a real-world scenario where a `logging` subproject chooses a logging backend implementation
and wants to make sure that all consumers will respect that choice. To do this, we can add a constraint on the `logging` project
that will reject all other logging backends.

When the `consumer` project depends on `logging`, it brings in the constraint, and the build fails.